Move services and login views into main/views
