+++ app/main/__init__.py
@@ -2,4 +2,5 @@ from flask import Blueprint
 
 main = Blueprint('main', __name__)
 
-from . import errors, services, login
+from . import errors
+from .views import services, login
